app/requirements.txt
autopep8==1.6.0
Flask==2.1.2
Flask-Session==0.4.0
mysql-connector-python==8.0.29
pip==22.1.2
requests==2.27.1